Fix asset reloading in development mode
The renderer was using a memoized class variable to cache the
concatinated react.js and components.js for server-side rendering. This
meant that when you updated a component, the server-rendered version would
not be refreshed.

Since #setup! is called every time a watched file is updated, it seems
like the natural place to reset the memozied string.