Fixes a gnarly test bug with server side rendering
Background: Running tests multiple times in a row could result in
errors, and could result in complete passing. In other words, the order
the tests ran in determined whether or not the tests would pass.

Ruby Test::Unit uses a psuedorandom seed to randomize the
order that tests are run in; you can specify a certain seed with
TESTOPS="--seed=xyz", which fixes the seed and thus the order the tests
will be run in.

The main source of the bug was that many of the tests rely on replacing
asset files and asserting that the results are as expected; however, the
asset pipeline uses mtime to determine when files have been changed.

*mtime's resolution is only 1 second, meaning that when a series of
actions took less than a second to run, the mtime wouldn't actually
change, and Rails asset pipeline wouldn't update the file list.*

This commit adds a hook to force reset the internal concatinated
react.js and components.js string. Another test requires a few calls to
sleep(1), because it is specifically testing that the asset pipeline
picks up changes to the file on disk correctly and renders them server
side.